Restructure __init__ files to be more user friendly
