+++ qctoolkit/__init__.py
@@ -1,7 +1,6 @@
 import typing
 
-__all__ = ["hardware", "pulses", "utils", "qcmatlab", "expressions", "serialization",
-           "MeasurementWindow", "ChannelID"]
+__all__ = ["MeasurementWindow", "ChannelID"]
 
 MeasurementWindow = typing.Tuple[str, float, float]
 ChannelID = typing.Union[str, int]
