+++ qctoolkit/hardware/__init__.py
@@ -1,5 +1,6 @@
-__all__ = [
-    'awgs',
-    'dacs',
-    'program'
-]
+from qctoolkit.hardware.setup import HardwareSetup
+
+from qctoolkit.hardware import awgs
+from qctoolkit.hardware import dacs
+
+__all__ = ["HardwareSetup", "awgs", "dacs"]
