+++ qctoolkit/pulses/__init__.py
@@ -1,12 +1,12 @@
-from qctoolkit.pulses.conditions import *
-from qctoolkit.pulses.function_pulse_template import *
-from qctoolkit.pulses.instructions import *
-from qctoolkit.pulses.loop_pulse_template import *
-from qctoolkit.pulses.multi_channel_pulse_template import *
-from qctoolkit.pulses.parameters import *
-from qctoolkit.pulses.pulse_template import *
-from qctoolkit.pulses.pulse_template_parameter_mapping import *
-from qctoolkit.pulses.repetition_pulse_template import *
-from qctoolkit.pulses.sequence_pulse_template import *
-from qctoolkit.pulses.sequencing import *
-from qctoolkit.pulses.table_pulse_template import *
+from qctoolkit.pulses.function_pulse_template import FunctionPulseTemplate as FunctionPT
+from qctoolkit.pulses.loop_pulse_template import ForLoopPulseTemplate as ForLoopPT
+from qctoolkit.pulses.multi_channel_pulse_template import AtomicMultiChannelPulseTemplate as AtomicMultiChannelPT
+from qctoolkit.pulses.pulse_template_parameter_mapping import MappingTemplate as MappingPT
+from qctoolkit.pulses.repetition_pulse_template import RepetitionPulseTemplate as RepetitionPT
+from qctoolkit.pulses.sequence_pulse_template import SequencePulseTemplate as SequencePT
+from qctoolkit.pulses.table_pulse_template import TablePulseTemplate as TablePT
+
+from qctoolkit.pulses.sequencing import Sequencer
+
+__all__ = ["FunctionPT", "ForLoopPT", "AtomicMultiChannelPT", "MappingPT", "RepetitionPT", "SequencePT", "TablePT",
+           "Sequencer"]
